Moves test classes into matching package.
Jira: OTPL-1828
